+++ projects/example/src/app/app.ts
@@ -41,6 +41,7 @@ export class App {
     { label: 'markdown viewer', link: '/pk-markdown-viewer' },
     { label: 'split', link: '/pk-split' },
     { label: 'modal', link: '/pk-modal' },
+    { label: 'otp input', link: '/pk-otp' },
     { label: 'pipes', link: '/pk-pipes' },
     { label: 'progress', link: '/pk-progress' },
     { label: 'qrcode', link: '/pk-qrcode' },