+++ projects/ngx-pk-ui/src/lib/pk-otp/pk-otp.model.ts
@@ -1,0 +1,2 @@
+export type PkOtpType = 'number' | 'char' | 'none';
+export type PkOtpSize = 'sm' | 'md' | 'lg';